--- a/cmd/tapio-sniffer/main.go
+++ b/cmd/tapio-sniffer/main.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/falseyair/tapio/pkg/unified"
+)
+
+func main() {
+	cli := unified.NewCLI()
+	
+	if err := cli.Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
+		os.Exit(1)
+	}
+}