… README.md
… docs/otel-enhanced-tracing.md
… ebpf/packet_analyzer.c
… ebpf/protocol_analyzer.c
… go.sum
… pkg/correlation/correlators.go
… pkg/correlation/engine_enhanced.go
… pkg/correlation/timeline.go
… pkg/correlation/types.go
… pkg/ebpf/enhanced_collector.go
… pkg/ebpf/enhanced_collector_stub.go
… pkg/ebpf/event_parsers.go
… pkg/ebpf/types.go
… pkg/journald/event_classifier.go
… pkg/journald/filters.go
… pkg/journald/parsers.go
… pkg/journald/pattern_matcher.go
… pkg/journald/reader.go
… pkg/performance/batch_processor.go
… pkg/performance/batch_processor_test.go
… pkg/performance/event_pipeline.go
… pkg/performance/object_pool.go
… pkg/performance/percpu_buffer.go
… pkg/performance/ring_buffer.go
… pkg/performance/ring_buffer_test.go
… pkg/performance_test.go
… pkg/resilience/load_shedding.go
… pkg/resilience/retry.go
… pkg/resilience/self_healing.go
… pkg/resilience/self_healing_test.go